Reverse maincoordinator
