+++ GarmentConfigurator/UI/Extensions/Shape+InsettableShape-Stroke.swift
@@ -1,0 +1,24 @@
+//
+//  Shape+InsettableShape-Stroke.swift
+//  GarmentConfigurator
+//
+//  Created by Aleksandr Shapovalov on 07/03/23.
+//
+
+import SwiftUI
+
+extension Shape {
+    func fill<Fill: ShapeStyle, Stroke: ShapeStyle>(_ fillStyle: Fill, strokeBorder strokeStyle: Stroke, lineWidth: Double = 1) -> some View {
+        self
+            .stroke(strokeStyle, lineWidth: lineWidth)
+            .background(self.fill(fillStyle))
+    }
+}
+
+extension InsettableShape {
+    func fill<Fill: ShapeStyle, Stroke: ShapeStyle>(_ fillStyle: Fill, strokeBorder strokeStyle: Stroke, lineWidth: Double = 1) -> some View {
+        self
+            .strokeBorder(strokeStyle, lineWidth: lineWidth)
+            .background(self.fill(fillStyle))
+    }
+}
